Add `spice pods` CLI command Also refactor API vs Manifest specs
